backend/requirements.txt
requests
fastapi
uvicorn[standard]
httpx
pydantic
python-dotenv
openai
pyyaml
